# Optimize server conversions for client-server array transfers