--- a/arkouda/pdarrayclass.py
+++ b/arkouda/pdarrayclass.py
@@ -1,12 +1,12 @@
 from __future__ import annotations
 from typing import cast, List, Sequence
 from typeguard import typechecked
-import json, struct
+import json
 import numpy as np # type: ignore
 from arkouda.client import generic_msg
 from arkouda.dtypes import dtype, DTypes, resolve_scalar_dtype, \
-     structDtypeCodes, translate_np_dtype, NUMBER_FORMAT_STRINGS, \
-     int_scalars, numeric_scalars, numpy_scalars
+     translate_np_dtype, NUMBER_FORMAT_STRINGS, \
+     int_scalars, numeric_scalars, numpy_scalars, get_server_byteorder
 from arkouda.dtypes import int64 as akint64
 from arkouda.dtypes import str_ as akstr_
 from arkouda.dtypes import bool as npbool
@@ -856,10 +856,13 @@ class pdarray:
         if len(rep_msg) != self.size*self.dtype.itemsize:
             raise RuntimeError("Expected {} bytes but received {}".\
                                format(self.size*self.dtype.itemsize, len(rep_msg)))
-        # The server sends us big-endian bytes so we need to account for that.
+        # The server sends us native-endian bytes so we need to account for that.
         # Since bytes are immutable, we need to copy the np array to be mutable
         dt = np.dtype(self.dtype)
-        dt = dt.newbyteorder('>')
+        if get_server_byteorder() == 'big':
+            dt = dt.newbyteorder('>')
+        else:
+            dt = dt.newbyteorder('<')
         return np.frombuffer(rep_msg, dt).copy()
 
     def to_cuda(self):
@@ -915,24 +918,8 @@ class pdarray:
             raise ModuleNotFoundError(('Numba is not enabled or installed and ' +
                                       'is required for GPU support.'))
 
-        # Total number of bytes in the array data
-        arraybytes = self.size * self.dtype.itemsize
-        
-        from arkouda.client import maxTransferBytes
-        # Guard against overflowing client memory
-        if arraybytes > maxTransferBytes:
-            raise RuntimeError(("Array exceeds allowed size for transfer. " +
-                               "Increase client.maxTransferBytes to allow"))
-        # The reply from the server will be a bytes object
-        rep_msg = generic_msg(cmd="tondarray", args="{}".format(self.name), recv_bytes=True)
-        # Make sure the received data has the expected length
-        if len(rep_msg) != self.size*self.dtype.itemsize:
-            raise RuntimeError("Expected {} bytes but received {}".\
-                               format(self.size*self.dtype.itemsize, len(rep_msg)))
-        # Use struct to interpret bytes as a big-endian numeric array
-        fmt = '>{:n}{}'.format(self.size, structDtypeCodes[self.dtype.name])
         # Return a numba devicendarray
-        return cuda.to_device(struct.unpack(fmt, rep_msg))
+        return cuda.to_device(self.to_ndarray())
 
     @typechecked
     def save(self, prefix_path : str, dataset : str='array', mode : str='truncate') -> str: