--- a/arkouda/pdarraycreation.py
+++ b/arkouda/pdarraycreation.py
@@ -3,9 +3,9 @@ import pandas as pd # type: ignore
 from typing import cast, Iterable, Optional, Union
 from typeguard import typechecked
 from arkouda.client import generic_msg
-from arkouda.dtypes import structDtypeCodes, NUMBER_FORMAT_STRINGS, float64, int64, \
+from arkouda.dtypes import NUMBER_FORMAT_STRINGS, float64, int64, \
      DTypes, isSupportedInt, isSupportedNumber, NumericDTypes, SeriesDTypes,\
-    int_scalars, numeric_scalars, get_byteorder
+    int_scalars, numeric_scalars, get_byteorder, get_server_byteorder
 from arkouda.dtypes import dtype as akdtype
 from arkouda.pdarrayclass import pdarray, create_pdarray
 from arkouda.strings import Strings
@@ -203,9 +203,11 @@ def array(a : Union[pdarray,np.ndarray, Iterable]) -> Union[pdarray, Strings]:
         raise RuntimeError(("Array exceeds allowed transfer size. Increase " +
                             "ak.maxTransferBytes to allow"))
     # Pack binary array data into a bytes object with a command header
-    # including the dtype and size. Note that the server expects big-endian so
-    # if we're using litle-endian swap the bytes before sending.
-    if get_byteorder(a.dtype) == '<':
+    # including the dtype and size. If the server has a different byteorder
+    # than our numpy array we need to swap to match since the server expects
+    # native endian bytes
+    if ((get_byteorder(a.dtype) == '<' and get_server_byteorder() == 'big') or
+        (get_byteorder(a.dtype) == '>' and get_server_byteorder() == 'little')):
         abytes = a.byteswap().tobytes()
     else:
         abytes = a.tobytes()