Improve the performance of `ak.array()` and `pdarray.to_ndarray()` by
optimizing how the server converts between bytes and pdarrays.
Previously, the server would write to and read from a big-endian memory
mapped file to convert between bytes and arrays but this is fairly
slow. Optimize this conversion by directly interpreting the underlying
memory as the type we're converting to. In the `ak.array()` case we
create the local array with `makeArrayFromPtr`, which will create an
array from the existing bytes without any copies. For the `to_ndarray()`
case `makeArrayFromPtr` is also used on some local memory, which is then
reinterpreted as bytes with `createBytesWithOwnedBuffer`.

Here's the performance improvement for 16-node-xc:

| config | to_ndarray | ak.array   |
| ------ | ---------: | ---------: |
| before |   33 MiB/s |   50 MiB/s |
| after  |  410 MiB/s |  175 MiB/s |

Where `ak.array()` is slower because there are more copies compared to
`to_ndarray()` on both the server and client side. Optimizing those
copies out is future work.

Part of #794